refactor/#9-home-ui : 홈,프로필수정 프래그먼트 -> 액티비티 수정
